requirements.txt
click==8.1.3
dataclasses-json==0.5.7
distro==1.8.0
intervaltree==3.1.0
libcst==0.4.9
marshmallow==3.19.0
marshmallow-enum==1.5.1
mypy-extensions==0.4.3
packaging==23.0
psutil==5.9.4
Pygments==2.14.0
pyre-check==0.9.17
pyre-extensions==0.0.30
python-magic==0.4.27
PyYAML==6.0
sortedcontainers==2.4.0
tabulate==0.9.0
TestSlide==2.7.0
typeguard==2.13.3
typing-inspect==0.8.0
typing_extensions==4.4.0
